fixed regex again fixed carinsurance bonus label visibility moved car Observable list from GUI to insuranceclass better menu button shadows Carinsurance calculations fixed editperson: insurance and incidentnumberrs now follows new person

+++ src/Incident/Incident.java
@@ -1,6 +1,7 @@
 package Incident;
 
 import Person.Person;
+
 import java.nio.file.Files;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -40,7 +41,7 @@ public abstract class Incident {
         this.timeOfIncident = timeOfIncident;
     }
 
-    public long getIncidentID() {
+    public int getIncidentID() {
         return incidentID;
     }
 }
